Updated project to use a vector of doubles instead of template values.
The project was updated from a template class because the template
had no compilable code which made it ineffective for using in a dll.
The member vector was changed from a template to be a type double
instead of a template type intended for numerical data. This ensures
the datatype used with the Population class is numeric (force double),
compilable code (.cpp file added), and all numeric values can be
updated to double value with no loss in precision.